# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: ComplyTime Platform API
  version: 0.1.0
  description: >
    REST surface served by `cmd/gateway`. Many `/api/*` routes register only when PostgreSQL
    is configured; `/api/config` is always available. After Google OAuth, the encrypted
    `studio_session` cookie authenticates `/api/*` requests (unless disabled). Optional
    `Authorization: Bearer <STUDIO_API_TOKEN>` bypasses the cookie for scripting.
    Security schemes below model bearer token and OAuth; cookie sessions are established via `/auth/*`.
servers:
  - url: http://localhost:8080
    description: Local gateway default (PORT=8080)
security:
  - bearerAuth: []
  - oauth2: [openid, email, profile]
...
paths:
...
  /api/audit-logs:
    get:
      tags: [AuditLogs]
      operationId: listAuditLogs
      summary: List audit logs for a policy
      parameters:
        - name: policy_id
          in: query
          required: true
          schema:
            type: string
        - name: start
          in: query
          description: Date YYYY-MM-DD
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'
        '400':
          description: Missing policy_id
    post:
      tags: [AuditLogs]
      operationId: createAuditLog
      summary: Store finalized audit log YAML
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditLogRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  audit_id:
                    type: string
        '400':
          description: Invalid YAML or JSON
  /api/audit-logs/{id}:
    get:
      tags: [AuditLogs]
      operationId: getAuditLog
      parameters:
        - $ref: '#/components/parameters/AuditId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLog'
        '404':
          description: Not found
  /api/audit-logs/promote:
    post:
      tags: [DraftAuditLogs]
      operationId: promoteDraftAuditLog
      summary: Promote draft audit log to official audit_logs (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [draft_id]
              properties:
                draft_id:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  draft_id:
                    type: string
        '400':
          description: Invalid body
        '409':
          description: Already promoted
        '403':
          description: Admin required
  /api/draft-audit-logs:
    get:
      tags: [DraftAuditLogs]
      operationId: listDraftAuditLogs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DraftAuditLog'
    post:
      tags: [DraftAuditLogs]
      operationId: createDraftAuditLog
      summary: Create a draft audit log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy_id, content, summary]
              properties:
                policy_id:
                  type: string
                content:
                  type: string
                summary:
                  type: string
                agent_reasoning:
                  type: string
                model:
                  type: string
                prompt_version:
                  type: string
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  draft_id:
                    type: string
  /api/draft-audit-logs/{id}:
    get:
      tags: [DraftAuditLogs]
      operationId: getDraftAuditLog
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftAuditLog'
        '404':
          description: Draft not found
    patch:
      tags: [DraftAuditLogs]
      operationId: patchDraftAuditLogEdits
      summary: Persist reviewer edits JSON map
      parameters:
        - $ref: '#/components/parameters/DraftId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reviewer_edits]
              properties:
                reviewer_edits:
                  type: object
                  additionalProperties:
                    $ref: '#/components/schemas/DraftReviewerEditEntry'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: saved
        '404':
          description: Draft not found
        '409':
          description: Draft already promoted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared secret from `STUDIO_API_TOKEN`; treated as admin-capable service identity when configured.
    oauth2:
      type: oauth2
      description: >
        Google OAuth 2.0 login flow used to mint the encrypted `studio_session` cookie.
        API requests authenticate via that cookie unless `bearerAuth` matches.
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Email
            profile: Profile
  parameters:
...
    AuditId:
      name: id
      in: path
      required: true
      schema:
        type: string
    DraftId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    AuditLog:
      type: object
      properties:
        audit_id:
          type: string
        policy_id:
          type: string
        audit_start:
          type: string
          format: date-time
        audit_end:
          type: string
          format: date-time
        framework:
          type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        content:
          type: string
        summary:
          type: string
        model:
          type: string
        prompt_version:
          type: string
    DraftAuditLog:
      type: object
      properties:
        draft_id:
          type: string
        policy_id:
          type: string
        audit_start:
          type: string
          format: date-time
        audit_end:
          type: string
          format: date-time
        framework:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
        content:
          type: string
        summary:
          type: string
        agent_reasoning:
          type: string
        model:
          type: string
        prompt_version:
          type: string
        reviewed_by:
          type: string
        promoted_at:
          type: string
          format: date-time
          nullable: true
        reviewer_edits:
          type: string
          description: JSON map serialized server-side
    DraftReviewerEditEntry:
      type: object
      properties:
        type_override:
          type: string
        note:
          type: string
    CreateAuditLogRequest:
      type: object
      required: [policy_id, content]
      properties:
        policy_id:
          type: string
        content:
          type: string
        model:
          type: string
        prompt_version:
          type: string
...